+++ xrlint/cli/engine.py
@@ -16,7 +16,7 @@ from xrlint.config import ConfigList
 from xrlint.config import get_core_config
 from xrlint.formatter import FormatterContext
 from xrlint.formatters import export_formatters
-from xrlint.linter import Linter
+from xrlint.linter import Linter, DatasetOpener
 from xrlint.plugin import Plugin
 from xrlint.result import Message
 from xrlint.result import Result
@@ -96,8 +96,6 @@ class CliEngine:
         for file_path, is_dir in get_files(self.files, global_filter):
             config = config_list.compute_config(file_path)
             if config is not None:
-                if isinstance(config.processor, str):
-                    config.processor.preprocess(file_path, config.opener_options or {})
                 linter = Linter(config=config)
                 result = linter.verify_dataset(file_path)
             else: