+++ xrlint/linter.py
@@ -7,7 +7,8 @@ from xrlint.config import Config
 from xrlint.config import get_core_config
 from xrlint.config import merge_configs
 from xrlint.constants import MISSING_DATASET_FILE_PATH
-from xrlint.result import Result
+from xrlint.processor import ProcessorOp
+from xrlint.result import Result, Message
 from xrlint.rule import RuleConfig
 from xrlint.rule import RuleOp
 
@@ -16,6 +17,7 @@ from xrlint._linter.rule_ctx_impl import RuleContextImpl
 
 # noinspection PyProtectedMember
 from xrlint._linter.verify_impl import verify_dataset
+from xrlint.util.formatting import format_message_type_of
 
 
 def new_linter(
@@ -97,44 +99,87 @@ class Linter:
         config = merge_configs(self._config, config)
         config = merge_configs(config, config_kwargs)
 
-        error: Exception | None = None
-        if not isinstance(dataset, xr.Dataset):
-            ds_source = dataset
-            dataset, error = open_dataset(ds_source, config.opener_options or {})
-            if not file_path and isinstance(ds_source, (str, Path)):
-                file_path = str(ds_source)
-
-        if dataset is not None and not file_path:
-            file_path = _get_file_path_for_dataset(dataset)
-
-        context = RuleContextImpl(config=config, dataset=dataset, file_path=file_path)
-
-        if error:
-            context.report(str(error), fatal=True)
-        elif not config.rules:
-            context.report("No rules configured or applicable.", fatal=True)
+        if isinstance(dataset, xr.Dataset):
+            file_path = file_path or _get_file_path_for_dataset(dataset)
+            messages = _verify_dataset(
+                config, dataset, file_path or _get_file_path_for_dataset(dataset)
+            )
         else:
-            for rule_id, rule_config in config.rules.items():
-                with context.use_state(rule_id=rule_id):
-                    _apply_rule(context, rule_id, rule_config)
+            file_path = file_path or str(dataset)
+            messages = _open_and_verify_dataset(config, dataset, file_path)
 
-        return Result.new(
-            config=config, file_path=context.file_path, messages=context.messages
+        return Result.new(config=config, file_path=file_path, messages=messages)
+
+
+def _open_and_verify_dataset(
+    config: Config, ds_source: Any, file_path: str
+) -> list[Message]:
+    opener_options = config.opener_options or {}
+    if isinstance(ds_source, str) and config.processor is not None:
+        processor_op = config.get_processor_op(config.processor)
+        try:
+            ds_path_list = processor_op.preprocess(ds_source, opener_options)
+        except (OSError, ValueError, TypeError) as e:
+            return [Message(message=str(e), fatal=True, severity=2)]
+        return processor_op.postprocess(
+            [_verify_dataset(config, ds, path) for ds, path in ds_path_list],
+            ds_source,
         )
+    else:
+        try:
+            dataset = open_dataset(ds_source, opener_options)
+        except (OSError, ValueError, TypeError) as e:
+            return [Message(message=str(e), fatal=True, severity=2)]
+        with dataset:
+            return _verify_dataset(config, dataset, file_path)
 
 
-def open_dataset(
-    file_path: str, opener_options: dict[str, Any] | None
-) -> tuple[xr.Dataset, None] | tuple[None, Exception]:
+def _verify_dataset(
+    config: Config,
+    dataset: xr.Dataset,
+    file_path: str,
+) -> list[Message]:
+    assert isinstance(config, Config)
+    assert isinstance(dataset, xr.Dataset)
+    assert isinstance(file_path, str)
+
+    context = RuleContextImpl(config=config, dataset=dataset, file_path=file_path)
+
+    if not config.rules:
+        context.report("No rules configured or applicable.", fatal=True)
+    else:
+        for rule_id, rule_config in config.rules.items():
+            with context.use_state(rule_id=rule_id):
+                _apply_rule(context, rule_id, rule_config)
+
+    return context.messages
+
+
+class DatasetOpener(ProcessorOp):
+
+    def preprocess(
+        self, file_path: str, opener_options: dict[str, Any]
+    ) -> list[tuple[xr.Dataset, str]]:
+        """Open a dataset."""
+        engine = opener_options.pop("engine", None)
+        if engine is None and file_path.endswith(".zarr"):
+            engine = "zarr"
+
+        dataset = xr.open_dataset(file_path, engine=engine, **(opener_options or {}))
+        return [(dataset, file_path)]
+
+    def postprocess(
+        self, messages: list[list[Message]], file_path: str
+    ) -> list[Message]:
+        return messages[0]
+
+
+def open_dataset(file_path: str, opener_options: dict[str, Any] | None) -> xr.Dataset:
     """Open a dataset."""
     engine = opener_options.pop("engine", None)
     if engine is None and file_path.endswith(".zarr"):
         engine = "zarr"
-
-    try:
-        return xr.open_dataset(file_path, engine=engine, **(opener_options or {})), None
-    except (OSError, TypeError, ValueError) as e:
-        return None, e
+    return xr.open_dataset(file_path, engine=engine, **(opener_options or {}))
 
 
 def _apply_rule(