+++ xrlint/plugin.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass, field
 from typing import Any, Type, Callable, Literal
 
 from xrlint.config import Config
-from xrlint.processor import Processor
+from xrlint.processor import Processor, ProcessorOp, register_processor
 from xrlint.rule import Rule, RuleOp, register_rule
 from xrlint.util.formatting import format_message_type_of
 from xrlint.util.importutil import import_value
@@ -34,8 +34,6 @@ class Plugin:
 
     processors: dict[str, Processor] = field(default_factory=dict)
     """A dictionary containing named processors.
-    
-    Processors are note yet supported.
     """
 
     @classmethod
@@ -49,8 +47,7 @@ class Plugin:
     def define_rule(
         self,
         name: str,
-        /,
-        version: str | None = None,
+        version: str = "0.0.0",
         schema: dict[str, Any] | list[dict[str, Any]] | bool | None = None,
         type: Literal["problem", "suggestion", "layout"] | None = None,
         description: str | None = None,
@@ -59,11 +56,24 @@ class Plugin:
     ) -> Callable[[Any], Type[RuleOp]] | None:
         return register_rule(
             self.rules,
-            name,
+            name=name,
             version=version,
             schema=schema,
             type=type,
             description=description,
             docs_url=docs_url,
+            op_class=op_class,
+        )
+
+    def define_processor(
+        self,
+        name: str,
+        version: str = "0.0.0",
+        op_class: Type[ProcessorOp] | None = None,
+    ):
+        return register_processor(
+            self.processors,
+            name=name,
+            version=version,
             op_class=op_class,
         )